requirements.txt
Flask
gunicorn
dnspython
idna
requests
requests-toolbelt